Audio alerts now working.
Not a huge change logic wise, but this did take some effort to get right since sending emails means nasty code. The major changes here are simply to update the emailer to make it check for the source parameter and then use the correct Solr core for the query. Some refactoring is also included, though the logic of the system is not changed.