Removes the code that converts an innocent number into an ascii representation.
This change is sure to bring smiles to many people's faces, as it solidifies our conversion to a system without oblique URLs, moving our system to one with less-oblique URLs. Now, instead of having urls with random ascii stuff in them, we'll at least have URLs with random numbers in them. More meaningful all around and less likely to be a cuss word as well. Can you believe that we currently have all four letter cuss words in our URLs up through ones starting with 'F3gz'? An exercise for the humble reader.