Big refactor to make source --> type.
This was needed because somehow I didn't realize until the final stages of the API that we had fields for source in our data and that we therefore couldn't use source as an indicator of oral arguments vs opinions. Long ago, when I first did this feature I wanted to use the word type, but avoided it because of the shadowing of the built in keyword. Now, with this second reason to use the most accurate term, I've switched back to using the word type.