Wraps up the API work, adding about a dozen new tests.
This is a bit of a big one, but one that's long overdue. You'll see that I've added a number of test_assets and a big test that iterates over them. The goal here is to catch and document any changes to the APIs. As an example, since these are versioned, you can already see the differences between different versions of the API by doing a simple diff. Pretty neat. On top of that, this will finally ensure that I don't do anything dumb to one API while tweaking another. Alas, there is a fair amount of complicated code in the API and keeping everything working the same as the underlying database changes isn't always easy!