Changes the commit model to be timer-based.
Currently we commit whenever we get something new, but we get too much content for that now. This change introduces a new management flag to the cl_update_index command that allows it to do a simple commit. The intention is that when this lands, we'll create a cron job that commits the index every five minutes.

The other thing this does is change the save() method for the Documents and Audio objects so they have a `force_commit` argument rather than a `commit` argument. This allows us to force a commit to the index, but we no longer do it by default.

 Finally, this reintroduces the countdown timers that seem to have gotten lost during a merge. These timers make it so the citations don't get found immediately when we add items to the index. Instead, they get found over the next 0-3600 seconds, as decided by a random int.