Adds a new stat to the homepage with upgrades to the naturalduration template filter.
"naturalduration" converts a number to either a dict of days, hours, minutes and seconds (months, years and milliseconds not supported) or to a colon-separated string as in the search results. This commit contains tests for this filter and upgrades to make it support dicts and weird values like 0, None or floats.