Updates the tests for audio so they work.
Introduces a couple neat new things. First, there's now the concept of both a SolrTestCase, and an IndexedSolrTestCase. The former sets up empty solr indexes for audio and opinions, and the latter subclasses the former to both set up those indexes and populate them with some items. SolrTestCase is useful if you just want to test something and need Solr to be set up. IndexedSortTestCase is useful if you want all the data in place and want to test at that point.

In addition to this, I've updated the audio tests and the search tests so more of those are working, and I've replaced our deepgetattr() function with a better one that supports default values. As usual, lots of stuff!