Replaces the current cited-by pages with search results instead.
This is a major new feature that we can enable now that we have the citation graph in Solr. Rather than seeing a fairly static list of opinions that cite the one you're looking at, you can now use all the standard Solr stuff to do the same. So, you can query for citations only from certain jurisdictions, or you can set up alerts for when there are new citations. The other great thing here is that since it uses pagination, feeds, all all the rest of the search infrastructure, this change meant the deletion of hundreds of lines of code, and the addition of only a few. Pretty great.