Uses a better method to disable the auto_now fields during the migration.
During migrations, we want to preserve the date created and date modified timestamps, but they're generated automatically. The previous technique that was in place was to remove the automatic generation of these fields, and to then put it back after the migration. Well, that works, but in the meantime, it's a required field and the fixtures fail if the field isn't provided, whereas they would otherwise be given a default value. Solution? Turn off the automatic values **only** during the migration by tweaking the models before we do the migration.