Rewrites the pagerank code to make it MUCH faster.
The old code was a bit bad. It iterated over Documents, and for each would pull the items that it cited, adding them to a networkx DiGraph. While that worked, a much better way to do it is to work directly with the citations between documents, ignoring the documents themselves completely. This version does exactly that, so instead of one query per document, there's one query, period. An improvement. It's also better organized, with relevant parts pulled out, has better comments, and is 20% shorter.