Updated to account for later paths to already visited nodes being possibly shorter than earlier paths.
For example, consider the following:

    Town of Greece
     │
     ├──> Elk Grove-──┐
     │       │        │
     │       V        V
     └──> Zelman-─-> Good News Club
                      │
                      └──> Rosenberger
                            │
                            └──> Marsh

If we get lucky, we'll go from Town of Greece to Elk Grove to Good News Club to Rosenberger to Marsh -- Four hops. But we might not be lucky, and instead we may go from Town of Greece to Elk Grove to Zelman to GNC to Rosenberger to Marsh. In a network with a max degree of separation == 4, we would only include GNC, Rosenberger and Marsh in the first scenario, not the second.

 So...if we only track nodes visited and don't keep track of whether we've found a shorter route since our first visit, we may omit these types of traversals. Ugh. This change is meant to begin tracking and checking whether we've found a shorter route. It's going to come at a performance cost though.