Light bulb: Shortest routes are found faster by going to the oldest cases first.
Consider this network:

   START ─> A ──> END
      │            ^
      └────────────┘

In this network, START cites to END directly and also with an intermediate node. END is older than the intermediate node, and so if we go to it first, we'll always discover the shortest routes more efficiently.

Boom.

("Boom" contingent upon not later realizing this was a horrible idea.)