Adds pagination to the visualization page.
Note that orphans is set to 2, so if you have 20, 21, or 22 items that will all display as page one. But if you have 23, page one will have only 20 items while page two has 3. This prevents people from going to the last page only to discover only one or two items.