Sets up Redis as the cache instead of LocMem.
I welcome discussion, but this is in place for both development machines and non. I debated this, but in the end, it makes sense to me for developers to have the same cache settings as the live site so they can more easily catch things that are cached but shouldn't be.

My next step will be to go through and try to knock out those areas of the code that shouldn't be cached.

Note that this configuration uses `PICKLE_VERSION: 1`. This is a workaround to make DRF work properly until they get pickle v2 working (the code is merged, but not yet on pip). Soon.