Adds a unique_together constraint and a migration to clean up old data.
This migration is designed to delete data, removing 3494 citation objects from the beta machine that were created during the migration from the live machine.

I fully expect this new constraint to cause problems as we do things, but it's necessary to enforce this so that our data will be good. As we encounter problems we can knock them out one by one.