API: Makes Opinion no longer nested inside Cluster.
I like this from a design perspective, but it sucks from a performance perspective, as you're literally pulling back a couple dozen opinions (and all their text and citations) when you look at a list of clusters. That is just nutso, and thus it must go away for now at least.