Field selection for all but the search endpoint.
Alas, the search endpoint is, and always will be a PITA since it doesn't rely on the usual Django models. Everything that's free in a model is a pain when it's backed by Solr. Anyway, I poked at making it support field selection, but it wasn't easy, so I gave up in order to focus on more important things.

Looks like the best way to do this will be by using the get(), fields() and get_fields() methods, but I could be wrong.