API: Adds usage logging using redis.
This is my first time using Redis, so hopefully it's all good, but more eyes on the new class in utils.py might be helpful. The basic idea here is to tally statistics about every request that's made against the API so that we can see who the big users are and so we can brag about how much it's being used. We do NOT tally how people are using the API, but we do tally that they use it. Anonymous users are tallied with the username of None. If we see too much usage there, we'll have to start looking into IP addresses and such.

 Because we don't want to slow down every API request with an additional increment command being sent to the database, we use Redis for this, which should be nearly instantaneous.

 For every request, we increment six keys:

 api:v3.count              --> The total number of times the API has been hit.
 api:v3.d:$date.count      --> The total number of times the API has been hit...today.
 api:v3.user:$userID.count         --> The total number of times the user has hit the API.
 api:v3.user:$userId.d:$date.count --> The total number of times the user has hit the API...today.
 api:v3.endpoint:$endpointName.count        --> The total number of times an endpoint has been hit.
 api:v3.endpoint:$endpointName.d$date.count --> The total number of times an endpoint has been hit...today.

 I believe with these six stats, we can do lots of simple and quick queries on the lookup side of things. A first example of this is in the view for the homepage, where we generate a total for how many queries in the last ten days.

The biggest question I had while working on this was which of Redis's data types to use, but in the end I kept it very simple using just keys with values.

 Anyway, more eyes welcome, but this mofo is landing.