UTILS: Adds a new function to simply return the full host with protocol and port.
Somehow this isn't built into Django, so you have to do wonky work to identify if the site is secure and if so use https, and if not use http, and the whole thing is annoying.

Enough. This provides a simple template tag {% get_full_host %}, throw it anywhere you want if all you're looking for is something like:

 https://www.courtlistener.com <-- On live
 http://127.0.0.1:8000 <-- dev
 http://courtlistener2.com:81 <-- testing