TESTS: Fixes test in donate.
We had a fun test that would only fail when UTC and PST were different days. For example, it's 9pm on the 4th PST, which is 5am on the 5th in UTC. PST has the day as the 4th while UTC has it as the 5th....and that was bad.