Applies a proximity query to the citator to make queries more accurate.
This is a particular problem when matching to opinions with multiple parallel citations, which previously would be returned even if the query spanned two citations.

For example, say you're looking up the citation 22 US 44. Well, prior to this fix, it would have been possible to get an item with these citations as a result:

> 22 F2d 33, 44 F3d 33, 99 US 99

Even though the 22 and the 44 and the US are all in different citations, it'd be a hit because it had all three tokens. By using proximity queries, we're able to ensure that the lookup only matches full citations because internally a multiValue Solr field is defined by placing 1000 empty tokens between each value.