Converts defendant search so it uses median, not average.
The old code used the average length of a case name, which was 67 chars when the code was written. The new version uses the median, which is 28 chars. Some cases have wickedly-long case names in our database, but we know that nobody uses those names when they write opinions. Instead, they use more concise ones. We shouldn't be looking so far back when we know the median is so much less. The median makes a lot more sense because it throws these danged things away.