Removes pure case name searching from the citation finder.
It already has a lot of safety checks, but they're just not getting the job done. There are still a few cases with such annoying case names that the citation finder was identifying them incorrectly *constantly*. The culprits have names like "State v. So", "The Jones", or "United States", which are mostly just stopwords for case names. Terrible.

Anyway, this is probably going to mean a lot fewer citations get identified, but it'll improve as we add more parallel citations...which we'll be doing soon.