New migrations for the database. More details and instructions below.
This removes the judges app from the codebase and cleans up the migrations by basically just deleting and recreating them. Migrations are represented in three ways. They're in files on disk (in the migrations directory of each app), their completion state is stored in the database in the `django_migrations` table, and the database itself is changed (through the addition and configuration of various tables and columns) if they are properly applied.

If you're working on master, this commit and the instructions below should align all three of these things, but it's likely that your database is in a different state than these instructions assume. Thus, if the instructions below don't work, and you're OK simply nuking your database I recommend doing that instead (manage.py reset_db).

But, assuming you like the data in your database, the approach here is to:

1. Delete all old migrations from your database:

    manage.py dbshell
    DELETE FROM django_migrations;

1. Fake all the migrations in this commit, including a few that shouldn't be faked. This updates the `django_migrations` table so it thinks the migrations have been applied:

    manage.py migrate --fake

1. Roll back a few of the faked migrations in the people_db app, to the last one that genuinely should have been faked:

    manage.py migrate people_db 0003 --fake

1. Apply the migrations that are included in this commit that need to be applied:

    manage.py migrate people_db