Rewrite of the ftm matching algorithm. Includes two approaches.
The first approach uses only the people DB and simply iterates over all the people in the DB, seeing if it can find a match. In testing, it didn't match very well, and had a very low hit rate.

The second approach uses the search engine and attempts to look up every entity in the FTM dataset. It has a higher hit rate.

In my next commit, I will remove the first approach, but this commit checks it in, in case we want to refer to it in the future.