Updates all pages to have the second navbar.
This is one of those commits that changes a lot of things. For the most part, you'll see that almost every page just got simpler. The exception is, of course, base.html, which gained a lot of complexity, but this is exactly how templates (and life) should work: Remove complexity from children; move it to parents.