Creates a new feature for maintenance mode.
The big change is that it is now triggered by a setting, not by a code change. The default settings, in 10-public.py do not enable maintenance mode, so when you want to enable it, all you have to do is add a line like this to your 11-private.py file:

MAINTENANCE_MODE_ENABLED = True

You have two options for allowing yourself through the maintenance mode warning. First, anybody that is "staff" is allowed through by default (though this can be disabled). Second, IP addresses can be added to the MAINTENANCE_MODE_ALLOWED_IPS list, and they will be allowed through as well.