Fixes a bunch of imports so they are done properly.
On the surface, this changes nothing. BUT if you ever override settings in a test, for example, that override won't work unless you properly import the settings from django.conf. Failing to do so will result in a lot of time being confused and annoyed. Thus, when I fix this one, I am fixing them all. May we never again make this mistake.