Turns off generateWordParts and generateNumberParts in Solr.
This is a very weird setting that makes it so that a phone number like 415-505-5555 will be searchable as 4155055555 or as any of the other tokens. That seems fine...at first, but it also means that a docket number like 1:10-cv-0022 will be searchable as 110-cv-0022, which is bad.

This won't take effect until we re-index, but it's a minor problem nobody has complained about, so, OK, fine.