Adds error handling and throttling.
The code here is complex, but the idea is to maximally scrape all of the PACER court websites simultaneously. So if the code is working properly, when you launch it, it will create a task for every PACER site, each of which will begin downloading a week of content. While that week is processing, the code will continually check if it is done. If so, it'll start the next week. If not, it will perform a linear backoff with a 30s cap. The code will also verify that it's not scraping into the future.

The net of this is that at any time we should have about 200 tasks queued/processing, one for each court. As each gets completed, more get added to the queue. It should be a very fast approach that won't flood the queue, but which will keep it well-supplied with work to complete.