Major performance improvement to fix docket indexing in #663.
This isn't a panacea — it needs more work, but it's a huge improvement nevertheless. The big change here is that when we index a docket instead of re-pulling the docket metadata (including atty's, parties, and firms) for every docket entry on the docket, we only pull it once across the entire thing. Not exactly a groundbreaking change, but a necessary one. In local testing, this changed a function from one that used 7GB of RAM and 10 minutes before crashing to one that used 2-3GB of RAM before finishing in about 2.5 minutes.

The next improvement that we'll need here is to figure out which docket entries changed when we receive a new docket upload. Right now if a docket has 30k items, we'll parse all of them and send all of them to Solr. Instead, we should figure out which have changed and only update those. Because of the way we de-normalize, changes to the docket itself (like the addition of a new party) will still force us to update every docket entry, but I'm not sure there's a ton we can do about that.