First pass at working APIs for docket and PDF upload, including functional models.
Whew, this took a lot of effing work. A major dependency was scrapers for all of the code that's in the old RECAP server, and that had to be done in Juriscraper. Outside of that, there's a fair bit of complexity here in terms of making the DRF work properly, and do what we expect. There are also a gazillion complex tests that are landing here, which is great.