Makes the free opinion scraper handle network issues properly.
This is a bigger change that I would have  hoped for. Ultimately, I had to rewrite the code almost completely. The problem was that there was no way in the old method for the calling script to learn that a PACER site was down and to react appropriately. In the new code, we do something pretty new: Keep track of Celery results. So, we'll send an asynchronous chain of tasks at a PACER website. The chain will go to the site, run the report, convert the report to useful objects in our DB, and then save the status to the DB, returning it in a result. If anything fails along the way, we'll have a result that says that the SCRAPE_FAILED, and we'll be able to move on.

The other tweak this makes is that it sets any old items that say they're in progress so that they say they're failed. This is a protection at the start of the code against some bug marking a court in progress for a prolonged period of time. Look, if a court says it's in progress 12 hours after we started it, that's a major issue, and it's clearly a failure. Let's treat it as such. (This can happen, by the way, if you forget to use a timeout in a request.get() call against a server that doesn't respond.)