Fixes #731 by adding a tag tool to RECAP documents.
This will allow us to create collections that can later be filtered down to. Tag names can be up to 50 characters in length and can either be very specific or totally random. Once this is in place and we have some nice collections built up, we can create a page listing interesting collections of PACER data.