Heavy optimization to the API including new tests to prevent regressions.
Focus here is on limiting the number of queries that the APIs produce. It's very easy to add a "+1" query to the API, which would produce one more query per object. We had a LOT of these before this commit, and now we don't have any outside of the people endpoints, which typically don't have performance issues.

I've also added tests which ensure that the number of queries produced by these endpoints doesn't regress. These will be annoying to maintain because they'll need to be updated whenever there's a new related object on an endpoint, but I think that's an acceptable price to pay for the protection the tests provide.