Adds the urls, view, and template re-organization for the party tab.
This is a pretty big change because it redoes the template hierarchy for the existing docket page. It used to be simply:

    base.html
      - view_docket.html

Now, it's:

    base.html
     - docket_tabs.html
       - view_docket.html

 ~~ and ~~

    base.html
     - docket_tabs.html
       - docket_parties.html

This allows both the parties page and the normal docket page to share a common base which has things like the docket metadata and the tab strip itself. This is pretty similar (though a bit more complicated) than how we handle the settings pages, which also have tabbed navigation.