fix(RECAP IA Parsers): Avoids updating the is_available field if it's already True
We currently are getting data from two sources. The first is the IA archive, which is still being
populated by people using the old RECAP extensions. The second place is the newer RECAP Extensions
themselves, which push to CourtListener. Our old code for parsing dockets assumed that if the
"available" field in the XML was blank, then the item wasn't available. Well, that's not true
anymore because now we get content from elsewhere. This fix tweaks the importer so that if the
is_available field is already True, we keep that value when we do an XML import.

A second piece of this will be needed on the server, where I'll update any item that has
is_available set to False, but has a document associated with it.

Closes #779