fix(RECAP API): Adds docket history support for "old" clients
Prior to 1.1.8, the RECAP Extensions sent docket history reports along with docket reports to the
same location. In 1.1.8, we changed this behavior so that docket history reports get sent with
their own upload_type. This si good, but alas, the older client lingers on and we need to handle
it in some form without it causing crashes.