fix(SCDB Import): Fixes import logic
This should have worked before, because this was supposed to be stable code, but there are two parts of the code that relied on a return value from self.set_if_falsy(). One part of the code assumed that it returned True if things came out OK, the other part assumed it returned True if there was an error. Obviously, both weren't true, so I rewrote it so that it returns True if things are OK, and false if there's an error.