feat(PACER Data): Adds crawler for pacer_doc_ids
pacer_doc_ids are available via a link we recently discovered called show_case_doc. The input for
the link is the document number, attachment number, and case number, and in return it gives you the
pacer_doc_id.

The code in this commit sets us up to get about 3.5M of those IDs that we're currently missing.
Partially addresses #774