fix(recap dockets): Uses show_case_doc links on dockets
This updates the link that we show to buy documents from docket pages so that if we
don't know the pacer_case_id for an item, we can still make a good link directly to
that item. The current implementation just takes you to the docket page if we don't
know the pacer_doc_id. This will be much better.