fix(PACER data crawler): Better restart methods
The current crawler doesn't have guaranteed ordering of the queries, and since a lot of
them fail, we're doomed to repeat the failing queries every time the script fails. This
change makes it possible for us to know via a log line what the last item was we did
(within 1000 items), and then to restart at around that spot.