feat(RECAP API): Adds cleanup for orphaned uploads
There are a handful of ways that people can currently upload a PDF prior to uploading
the docket it belongs to. We try to mitigate the damage this causes by retrying document
uploads a handful of times over the course of an hour or so. That helps, but we need a
bigger approach to cleaning up these orphaned documents, which this fix should do.

This fix makes it so that as we add new documents to a docket, we keep track of any that
we created. Once the docket is done, we check the processing queue for any failed items
for that docket that may exist. If we find any, we reprocess them, hopefully adding them
to the docket. This means that when the docket is first uploaded by somebody, they can
really unlock a lot of content that was previously lingering in the processing queue as
failed uploads.