fix(RECAP API): Fix for missing pacer_case_ids on att pages
Attachment pages aren't getting submitted with good pacer_case_id values. I've fixed the
existing content, and I will soon fix the client so we don't get more of it in new versions
of RECAP. This makes sure that if we *do* continue to get bad data, we'll process is nicely.

This also fixes a problem with creating open file handles. I don't think it caused any
issues, but it's better to fix it.

Addresses freelawproject/recap#226