feat(RECAP API/Data): Removes unique constraint on pacer_doc_id
This addresses issue #765, in which we identified that the same document, with the same
pacer_doc_id, can appear in multiple cases. This happens a lot, but until now when this
happened, we would simply accept the first such document, and latter ones would be
rejected. That's problematic. This change removes the unique constraint on pacer_doc_id,
and makes it so that we can accept the same document multiple times over. What this does
*not* do is provide any kind of linkage or aliasing or anything of that sort so that such
duplicated documents can be uploaded to the Archive only once. Instead, they have to get
re-purchased, and uploaded once per instance. This isn't great, but the complexity of
providing aliases or anything like that is pretty high with our current data model. It's
not crazy to consider that we may want to fix it another way later, but for now at least
this will mean accepting a lot more documents, which we currently reject.

Like many issues I'm fixing right now, fixing this one will benefit from re-parsing all
of our existing dockets, which may have rejected documents. Once that's done, it's probably
worth reparsing all of our failed documents as well. So it goes.

Fixes #765.