fix(RECAP APIs): Catches all exceptions from orphan re-processing
This uses the rare, "bare except" statement, but that's OK because this is an instance where
we've already caught the exception in the past and hopefully dealt with it. The time we
caught it in the past was the first time that the item was processed. No need to catch and
handle the exceptions again now.