fix(RECAP API): Avoids two cases where orphaned PDFs get created
When the debug flag is active on a RECAP upload, we shouldn't be saving content over to the RECAP
storage area. Similarly, when we are unable to save to the DB, we should avoid leaving files in
the RECAP storage area.

The problem caused by both of these issues (aside from wasting storage) is that when the file is
successfully uploaded, it gets a _1 appended to it, because the testing or the failed file is
still sitting in the storage area. That's not good!