fix(RECAP API): Ensures that cases are properly split by pacer_case_id
This fix is related to freelawproject/recap#208, which deals with the fact that a docket can have
multiple pacer_case_ids. Right now we handle that pretty slipshod. Some of the time we merge such
cases, and other times we establish multiple dockets.

The correct thing to do, long term, is to allow each docket to have multiple pacer_case_id values.
That way, one docket sheet would show all of the related documents in the case. This accomplishes
the opposite: It enforces separation of the content into separate dockets. So, one docket number can
and will have multiple docket objects in CL and RECAP. That's bad, BUT until we can update our data
model and fix freelawproject/recap#208, we should at least be internally consistent about our own
data model, which is what this does.

Prior to this change, you'd merge things some of the time and not other times, which caused
headaches when trying to upload PDFs. Not good.