fix(search): Fixes bulk search indexing for some docket entries of RECAP dockets
Prior to this, if we had a docket entry that lacked a sub document, we wouldn't index that docket
entry. That was bad, and I'm about to fix it manually on the server. This only caused issues when
we indexed as dockets, but not which we only did during reindexing. In day-to-day usage, we index at
the document-level, which this bug didn't affect.

Closes #784