fix(search): Fixes bulk search indexing for some docket entries of RECAP dockets
This makes it so that minute entries are not searchable. I'm not sure what a better solution would
look like technically. The issue is that in Solr we need an ID for every item, and the one we use is
the ID from the document itself. Since we lack a document for these items, we lack an ID, and we're
a bit stuck. Our ID field in Solr can accept negative values, so one fix could be to use the
negative number space as an ID for these, but that is quite a hack. Another fix would be to use
a float field for these, say, and to use the docket entry number + the document number + the
attachment number as the ID. Maybe that'd work too. But for now, easiest thing is to move forward
and just not index these values.

Related to #784