fix(free opinion scraper): Fixes design logic error that caused no scraping.
The issue here was that we logged the last day that things were scraped, say it's Christmas, and
then we used that date as the marker for what was done. The assumption was that if we scraped on
Christmas, then we had all the content from that day, and the next scraper could pick up anything
*after* that day. Well, folks, let's say the cron job scrapes at 1am on Christmas. If that's the
case, let me tell you, you didn't get the opinions for Christmas, as your next scrape should use
Christmas as its starting point, NOT the next day.

So...this fixes that. Since we scrape just after midnight, this change ensures that our scraper does
not mark the day of last scrape as being complete. Yuck, nasty stuff.

Fixes: https://github.com/freelawproject/juriscraper/issues/199