fix(free opinion scraper): Makes retry mechanism less patient
The old way would try 20 times with a 10s delay between each try and a 300s timeout on the server,
for a maximum retry period of 21 × (10s + 300s) = 6600s or about 150 minutes. That's crazy.

This changes it to 2 retries with a with 5s in between and a 300s timeout for a total of
3 × (5s + 300s) or about 20 minutes. That's still a lot, but it's much less crazy.