views.py: Content-Disposition inline for PDF downloads
Most users tend to want PDFs displayed in their browsers.

Those who want otherwise can override (such as me).

But setting Content-Disposition: attachment does not allow most users
to see PDFs in their browers when they want to. It is undesirable and
counterproductive.

Possibly a better fix is to not specify Content-Disposition at all,
but I sat on this for 2 months without doign anything with that, so
let's go ahead with this.

Fixes #777.