feat(recap api): Adds parsers for docket history reports
This turned out to be pretty easy since it's 95% the same code as in the docket endpoint. The only
thing that makes this commit interesting is that in order to avoid repeating a bunch of code, the
docket endpoint is refactored considerably so that code could be used in the docket history report.

Other than that, this is fairly straightforward stuff.