fix(contact form): Makes the form weird in the hope of better defeating spam.
For the past week, we've been getting a lot of random and annoying spam from the contact form. My
hope is that by renaming the subject field as the phone_number field, and then by nuking messages
that start with digits, we can stop the spam. We'll see how well it works.

h/t @waldoj