simple_pages/urls.py: Remove incorrect mime_type claim
Don't generate the static file regexp from a list of strings that is
otherwise unused -- just put the list in the regexp directly. It's
shorter and more readable and places the information at the point of
need.

Also, don't call them MIME types when they aren't (confusing!) --
they're leading path strings, or whatever.