fix(parties): Removes unique constraint on attorneys
This removes the unique constraint on attorneys so that we can have the same attorney repeated on
different dockets. This is a *major* change in the way we conceptualize attorneys. Instead of trying
to normalize attorneys across dockets, we're going to leave that kind of thing up to others
downstream of us. Before now, we tried to match up attorneys across dockets by using their name and
address. This worked well! But it's riddled with problems. For example, the address information for
the same real-life attorney can be different on two different dockets. How do we account for that?
We could model it, but if we do that, how do we know which is right? How do we know which is most up
to date? How do we lookup attorneys by address? Is it worth it? What do we get out of it? Etc.
It just gets nasty.

So. I'm going to blow up our database and make it so that the same real-life attorney is
represented by many objects in the DB. This will preserve accuracy (each docket has its own copy of
the attorney to monkey with), and will simplify our merging algorithms because they'll only have to
deal with data on a single docket at a time. OTOH, it demands a recrawl of ALL the PACER/RECAP data,
which will likely take weeks, and it will require downstream folks to do this kind of de-duping on
their dime — something I'd hoped to save them from.

We currently have about 600k attorneys in the DB having about 5.3M roles for the parties they
represent. This number is about to go *way* up.

So it goes. Big changes afoot, my friends.

Partially addresses #755.