fix(party page): Moves extra_info to the PartyType field and removes date_sourced from attys.
Moving extra_info is needed because it turns out that the extra info is related to the docket,
not the party, so that move needs to happen. For now this field will not be populated by the old
data, but the parsers will begin filling it in for the new data. The next step is to re-run the
parsers, which will fix this for all the old data too.

Removing date_sourced makes sense because attorneys are no longer shared across dockets. That means
that we eliminate the need to keep track of the latest information. Whatever is on the docket is
the best stuff we've got.