fix(party page): Major rewrite of attorney/party parsing.
The big changes here:

 - date_sourced is no longer used (see previous commit).
 - date_terminated and extra_info are moved to the PartyType
 - We do smart things (hopefully) when we do/don't have/receive terminated parties (see
   disassociate_extraneous_entities).
 - Transactions are used to ensure that parties either work properly or fail completely.
 - Attorney look ups are focused within the current docket only and no longer attempt to merge
   across dockets.

In general, this should result in much better PACER entity parsing. I wrote numerous additional
tests too, but still I feel uncomfortable with this because it's so big and because I didn't break
it down into smaller pieces that could be deployed/tested properly. It's a bit late for that now,
but the good news is that broken RECAP parsing can be fixed without causing too much damage or
downtime and I *think* this should go smoothly.

The next step is to begin reparsing the parties section for everything we've got. In chronological
order. Oye.