fix(scdb merging): Tweaks to SCDB merging to make it work better on corner cases
First corner case: When there's no docket number in the matched item in the database. In that case,
we winnow based on something else instead of assuming it's no good.

Second corner case: An error during parsing of federal citations. In this case, the error is still
thrown, but the log is improved to show the value that threw the error.